chore: fix imports after renaming handler file
